pyproject.toml
[project]
name = "assemblyai-challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.15"
dependencies = [
    "fastapi>=0.116.1",
    "pydub>=0.25.1",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
    "eventlet>=0.40.2",
    "uvloop>=0.21.0",
    "orjson>=3.11.1",
    "loguru>=0.7.3",
    "sse-starlette>=3.0.2",
    "yt-dlp[browser]>=2025.7.21",
    "requests>=2.32.4",
    "python-dotenv>=1.1.1",
    "assemblyai[extras]>=0.42.0",
    "groq>=0.30.0",
    "aioredis>=2.0.1",
    "cbase64>=0.0.9",
]
